Migrate status reporting to molecular methods
